✨ migrate IdentityMint store to pinia
